Significant revision of proposal.
The previous iteration of my proposal suggested an X.Y.Z form in which
the Z position incremented when no other position increments, implying
an untested release the average user would want to avoid; this was
completely incompatible with the traditional use of this position as
one we're supposed to be able to safely update to (regardless of how
often this turns out to be false...). The new proposal dedicates this
position to bug-fix releases and adds a -D or DEV position for
releasing incremental work.